split Evaluator.cpp
